Refractored App.js to PostsPage.js component to make room for a login system
