+++ instagram/src/components/authentication/authenticate.js
@@ -1,0 +1,10 @@
+import React, { Component } from "react";
+
+const authenticate = Comp =>
+  class extends Component {
+    render() {
+      return <Comp />;
+    }
+  };
+
+export default authenticate;
